twister: move runner classes out
Move runners into own file.
